requirements.txt
langchain
langchain_openai
langchain_core
langchain_community
langchain-groq

langserve
fastapi
uvicorn
sse_starlette
streamlit

python-dotenv

chromadb
faiss-cpu
sentence-transformers

pypdf
PyPDF2
beautifulsoup4
bs4

wikipedia
arxiv
langchainhub
groq
sse_starlette